Move MathFormats into its own PSR-4 compliant file
MathFormats was declared as the second class in src/Math/Math.php, so
it was not reachable by name through composer.json's "SRF\\": "src/"
PSR-4 mapping, which resolves SRF\Math\MathFormats to the nonexistent
src/Math/MathFormats.php. Nothing else registered it: extension.json's
AutoloadClasses covers only SRF\Carousel, AutoloadNamespaces only
SRF\DataTables\, and composer's classmap only formats/. Composer's
optimized autoloader does not help either, since its ClassMap
generator skips PSR-4-non-compliant classes.

This never surfaced in production because MathFormats is referenced
only from inside Math::getResultText(), by which point Math.php has
already been loaded as a side effect of autoloading SRF\Math\Math.

It does surface for the unit tests added in this branch, which name
the class directly. They passed only because PHPUnit sorts discovered
files and tests/phpunit/Unit/Formats/MathTest.php runs first,
instantiating SRF\Math\Math and defining MathFormats as a side effect.
Run on their own the whole file errored:

  composer phpunit -- --filter MathFormatsTest
    before: Tests: 28, Errors: 28
            Error: Class "SRF\Math\MathFormats" not found
    after:  OK (28 tests, 28 assertions)

Moving the class to src/Math/MathFormats.php makes it autoloadable on
its own, so the tests no longer depend on an unrelated test running
first. The .phan/baseline.php entry for medianFunction moves with it,
since baseline keys are file paths.
